passed second test case. Was able to create a NPS

+++ src/com/mycompany/a1/MovableObject/MissileLauncher.java
@@ -5,7 +5,7 @@ import com.mycompany.a1.GameWorldObjects.MovableObject;
 
 public class MissileLauncher extends MovableObject{
 	
-	public MissileLauncher(int x, int y, int heading, int speed) {
+	public MissileLauncher(float x, float y, int heading, int speed) {
 		super.setColor(ColorUtil.YELLOW);
 		super.setX(x);
 		super.setY(y);
